+++ src/components/main/Header/Logo/style.ts
@@ -5,6 +5,5 @@ export const Logo = styled(Link)`
   color: white;
   font-size: 22px;
   font-weight: 800;
-  font-family: Gilroy;
   text-decoration: none;
 `;
