+++ src/lib/redux/User/UserSlice.ts
@@ -1,6 +1,7 @@
 import { createSlice } from "@reduxjs/toolkit";
+import { useDispatch } from "react-redux";
 
-export const userSlice = createSlice({
+export const UserSlice = createSlice({
   name: "user",
   initialState: {
     token: null,
@@ -18,6 +19,27 @@ export const userSlice = createSlice({
 export const {
   loginSuccess: loginSuccessActionCreator,
   logoutSuccess: logoutSuccessActionCreator,
-} = userSlice.actions;
+} = UserSlice.actions;
 
-export default userSlice.reducer;
+export function loadLocalUser() {
+  const dispatch = useDispatch();
+  try {
+    const user = localStorage.getItem("loginUser");
+    if (!user) return;
+    dispatch(loginSuccessActionCreator(user));
+  } catch (e) {
+    console.log("localStorage is not working");
+  }
+}
+export function loadSessionUser() {
+  const dispatch = useDispatch();
+  try {
+    const user = sessionStorage.getItem("loginUser");
+    if (!user) return;
+    dispatch(loginSuccessActionCreator(user));
+  } catch (e) {
+    console.log("sessionStorage is not working");
+  }
+}
+
+export default UserSlice.reducer;
